requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
psycopg2-binary==2.9.3
python-dotenv==0.19.0
requests==2.26.0
textblob==0.15.3
wikipedia-api==0.5.4
newspaper3k==0.2.8
duckduckgo-search==3.9.3
python-dateutil==2.8.2
nltk==3.8.1
numpy==1.21.0
psutil==5.9.5
flask-cors==3.0.10
openai==0.27.8
